% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma2eigen}
\alias{arma2eigen}
\title{Convert Arma matrix to Eigen matrix}
\usage{
arma2eigen(arma_A)
}
\arguments{
\item{arma_A}{A matrix.}
}
\description{
Convert Arma matrix to Eigen matrix
}
\author{
Andrew DiLernia
}
