% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockBoot_cpp}
\alias{blockBoot_cpp}
\title{Moving Block Bootstrap}
\usage{
blockBoot_cpp(mvts, winLength, nBoots = 500L, stationary = FALSE)
}
\arguments{
\item{mvts}{\eqn{n} x \eqn{p} matrix of observed \eqn{p}-variate time series}

\item{winLength}{nonnegative window length parameter}

\item{nBoots}{Number of bootstrap samples}

\item{stationary}{Logical value indicating whether to use a variable window length block bootstrap (TRUE) as described by Politis & Romano (1994), or a fixed window length block bootstrap (FALSE) as described by Kunsch (1989)}
}
\value{
3D array of dimension \eqn{n} x \eqn{p} x nBoots containing the nBoots bootstrap samples
}
\description{
This function implements the moving block bootstrap as proposed by Kunsch (1989).
}
\references{
Kunsch, H. R. (1989). "The Jackknife and the Bootstrap for General Stationary Observations". Annals of Statistics. 17 (3): 1217-1241. doi:10.1214/aos/1176347265.

Politis, D. N.; Romano, J. P. (1994). "The Stationary Bootstrap". Journal of the American Statistical Association. 89 (428): 1303-1313. doi:10.1080/01621459.1994.10476870. hdl:10983/25607.
}
\author{
Andrew DiLernia
}
