% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correlations_covariance_hedges_olkin_cpp}
\alias{correlations_covariance_hedges_olkin_cpp}
\title{Asymptotic Covariance Matrix for Correlations of Hedges and Olkin (1983)}
\usage{
correlations_covariance_hedges_olkin_cpp(
  correlation_matrix,
  n,
  correlation_pairs,
  structure = "unstructured"
)
}
\arguments{
\item{correlation_matrix}{\eqn{p} x \eqn{p} correlation matrix}

\item{n}{number of observations}

\item{correlation_pairs}{(choose(\eqn{q}, 2) + \eqn{q}) x 4 matrix of indices for correlations equal to royVarhelper(p, errors = FALSE) where \eqn{q} = \eqn{p} choose \eqn{2}.}

\item{structure}{Optional covariance structure, indicating whether to use a covariance estimator in which every entry is estimated ("unstructured"), covariances between correlations with disjoint pairs of variables are set to 0 while others are estimated ("intersecting-pairs"), or off-diagonal entries are set to 0 and diagonal entries are estimated ("diagonal").}
}
\value{
\eqn{q} x \eqn{q} covariance matrix
}
\description{
This function calculates an asymptotic estimate of the covariance matrix for a vector of correlations as given by Hedges and Olkin (1983).
}
\references{
Hedges, L. V. and Olkin, I. (1983). Joint distributions of some indices based on correlation coefficients. In Studies in Econometrics, Time Series, and Multivariate Statistics, 437–454. Academic Press.
}
\author{
Andrew DiLernia
}
