% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corrMat_cpp}
\alias{corrMat_cpp}
\title{Calculate marginal or partial correlation matrix}
\usage{
corrMat_cpp(tsData, partial = TRUE)
}
\arguments{
\item{tsData}{An n x p data matrix.}

\item{partial}{Logical. Whether to calculate partial (TRUE) or marginal (FALSE) correlation matrix}
}
\value{
p x p correlation matrix
}
\description{
Calculate marginal or partial correlation matrix
}
\author{
Andrew DiLernia
}
