% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosTaper_cpp}
\alias{cosTaper_cpp}
\title{Function for Hann window taper}
\usage{
cosTaper_cpp(u)
}
\arguments{
\item{u}{Vector of indices for window}
}
\description{
Function for Hann window taper
}
\author{
Andrew DiLernia
}
