% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov2corr_cpp}
\alias{cov2corr_cpp}
\title{Convert covariance matrix to correlation matrix}
\usage{
cov2corr_cpp(cmat)
}
\arguments{
\item{cmat}{A covariance matrix.}
}
\description{
Convert covariance matrix to correlation matrix
}
\author{
Andrew DiLernia
}
