% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossCov_cpp}
\alias{crossCov_cpp}
\title{Calculate cross-covariance of two vectors}
\usage{
crossCov_cpp(u, ts1, ts2)
}
\arguments{
\item{u}{Vector of indices for window}

\item{ts1}{First time series vector}

\item{ts2}{Second time series vector}
}
\description{
Calculate cross-covariance of two vectors
}
\author{
Andrew DiLernia
}
