% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigen2arma}
\alias{eigen2arma}
\title{Convert Eigen matrix to Arma matrix}
\usage{
eigen2arma(eigen_A)
}
\arguments{
\item{eigen_A}{A matrix.}
}
\description{
Convert Eigen matrix to Arma matrix
}
\author{
Andrew DiLernia
}
