% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMult2}
\alias{eigenMult2}
\title{Multiply two matrices}
\usage{
eigenMult2(A, B)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}
}
\description{
Multiply two matrices
}
\author{
Andrew DiLernia
}
