% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMult3}
\alias{eigenMult3}
\title{Multiply three matrices}
\usage{
eigenMult3(A, B, C)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}

\item{C}{A matrix.}
}
\description{
Multiply three matrices
}
\author{
Andrew DiLernia
}
