% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMult4}
\alias{eigenMult4}
\title{Multiply four matrices}
\usage{
eigenMult4(A, B, C, D)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}

\item{C}{A matrix.}

\item{D}{A matrix.}
}
\description{
Multiply four matrices
}
\author{
Andrew DiLernia
}
