% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expTaper_cpp}
\alias{expTaper_cpp}
\title{Function for exponential window taper}
\usage{
expTaper_cpp(u)
}
\arguments{
\item{u}{Vector of indices for window}
}
\description{
Function for exponential window taper
}
\author{
Andrew DiLernia
}
