% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcCovHelper.R
\name{invCov2part}
\alias{invCov2part}
\title{Partial Correlation Matrix}
\usage{
invCov2part(icmat)
}
\arguments{
\item{icmat}{\eqn{p} x \eqn{p} inverse covariance matrix}
}
\value{
\eqn{p} x \eqn{p} partial correlation matrix
}
\description{
This function uses the inverse covariance matrix to calculate the partial correlation matrix
}
\examples{
# Generate multivariate time series with 5 variables from a
# first-order VAR model with 50 time points
set.seed(1994)
myTS <- varSim(nt = 50, coeffMat = diag(0.50, 5), covMat = diag(1, 5))

# Calculate partial correlation matrix
invCov2part(solve(cov(myTS)))

}
\author{
Andrew DiLernia
}
