% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invCov2part_cpp}
\alias{invCov2part_cpp}
\title{Convert inverse-covariance matrix to partial correlation matrix}
\usage{
invCov2part_cpp(icmat)
}
\arguments{
\item{icmat}{An inverse-covariance matrix.}
}
\description{
Convert inverse-covariance matrix to partial correlation matrix
}
\author{
Andrew DiLernia
}
