% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mkSymm_cpp}
\alias{mkSymm_cpp}
\title{Adjust matrix to be symmetric}
\usage{
mkSymm_cpp(myMat)
}
\arguments{
\item{myMat}{A matrix.}
}
\description{
Adjust matrix to be symmetric
}
\author{
Andrew DiLernia
}
