% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalFuns.R
\name{multiBlockBoot}
\alias{multiBlockBoot}
\title{Multi-Subject Block-Bootstrap for Multivariate Time Series}
\usage{
multiBlockBoot(mvts, winLength = NULL, nboots = 500)
}
\arguments{
\item{mvts}{List of \eqn{K} number of \eqn{nt} x \eqn{p} matrices of the observed \eqn{p}-variate time series.}

\item{winLength}{Specified window length for block-bootstrap (Optional). If not specified, block-length is determined using the method described in Patton, Politis and White (2009).}

\item{nboots}{Number of bootstrap samples.}
}
\value{
A list of \eqn{K} arrays containing block-bootstrap samples.
}
\description{
This function obtains block-bootstrap samples for multi-subject time series data.
}
\examples{
# Generate multivariate time series for 10 subjects with 5 variables from a
# first-order VAR model with 100 time points
set.seed(1994)
multiSubjTS <- lapply(1:10, FUN = function(subj) {pcCov::varSim(nt = 100, coeffMat = diag(0.50, 5), covMat = diag(1, 5))})

# Obtain 1000 block-bootstrap samples for each subject
bootRes <- multiBlockBoot(mvts = multiSubjTS, nboots = 1000)

}
\references{
Politis, D.N. and H. White (2004), Automatic block-length selection for the dependent bootstrap, \emph{Econometric Reviews}, 23(1), 53-70.
}
\author{
Andrew DiLernia
}
