% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiBlockBootCorr_cpp}
\alias{multiBlockBootCorr_cpp}
\title{Multi-Subject Moving Block Bootstrap for Correlation Coefficients}
\usage{
multiBlockBootCorr_cpp(
  mvts,
  winLengths,
  nBoots = 500L,
  stationary = FALSE,
  partial = TRUE
)
}
\arguments{
\item{mvts}{3D array of dimension \eqn{n} x \eqn{p} x \eqn{K} of \eqn{n}-length observed \eqn{p}-variate time series for \eqn{K} individuals}

\item{nBoots}{Number of bootstrap samples}

\item{stationary}{Logical value indicating whether to use a variable window length block bootstrap (TRUE) as described by Politis & Romano (1994), or a fixed window length block bootstrap (FALSE) as described by Kunsch (1989)}

\item{partial}{Logical value indicating whether to implement block bootstrap for the partial (TRUE) or marginal (FALSE) correlation coefficients}

\item{winLength}{nonnegative window length parameter}
}
\value{
3D array of dimension \eqn{q} x nBoots x \eqn{K} containing the \eqn{q=} choose(\eqn{p}, 2) correlations for each of the nBoots bootstrap samples for each of the \eqn{K} individuals
}
\description{
This function implements the moving block bootstrap as proposed by Kunsch (1989) for the correlation coefficients of a multi-subject data set.
}
\references{
Kunsch, H. R. (1989). "The Jackknife and the Bootstrap for General Stationary Observations". Annals of Statistics. 17 (3): 1217-1241. doi:10.1214/aos/1176347265.

Politis, D. N.; Romano, J. P. (1994). "The Stationary Bootstrap". Journal of the American Statistical Association. 89 (428): 1303-1313. doi:10.1080/01621459.1994.10476870. hdl:10983/25607.
}
\author{
Andrew DiLernia
}
