% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{partialCov_cpp2}
\alias{partialCov_cpp2}
\title{Taylor Series Estimate of Covariance Matrix for Partial Correlations}
\usage{
partialCov_cpp2(ts, bw, iMatq, iMate, q, resids)
}
\arguments{
\item{ts}{\eqn{nt} x \eqn{p} matrix of observed \eqn{p}-variate time series.}

\item{bw}{nonnegative bandwidth parameter.}

\item{iMatq}{matrix of indices for partial correlations equal to unique(royVarhelper(p)[, 1:2]).}

\item{iMate}{matrix of indices for partial correlations equal to royVarhelper(p, errors = TRUE).}

\item{q}{number of unique partial correlations equal to choose(\eqn{p}, 2).}

\item{resids}{\eqn{nt} x \eqn{2q} matrix of regression residuals (optional). Default is to use ordinary least squares to calculate the residuals.}
}
\value{
\eqn{q} x \eqn{q} covariance matrix
}
\description{
This function calculates a second-order Taylor Series estimate of the covariance matrix for partial correlations of a weakly stationary multivariate time series.
}
\author{
Andrew DiLernia
}
