% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{royVar_cpp}
\alias{royVar_cpp}
\title{Roy Asymptotic Variance}
\usage{
royVar_cpp(iMat, tsData, q, bw = 10L)
}
\arguments{
\item{iMat}{Matrix of correlation indices}

\item{tsData}{Matrix of observed n-length p-variate time series}

\item{q}{Integer equal to the number of unique variables pairs given by choose(p, 2)}

\item{bw}{Bandwidth parameter}
}
\description{
Roy Asymptotic Variance
}
\author{
Andrew DiLernia
}
