% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{royVcm2_cpp}
\alias{royVcm2_cpp}
\title{Variance Components Model}
\usage{
royVcm2_cpp(
  rs,
  sigmas,
  sigEigs,
  sigMean,
  delta = 0.001,
  maxIters = 100L,
  sig0 = 0.1
)
}
\arguments{
\item{rs}{column vector containing q x K unique correlations.}

\item{sigmas}{3D array of K estimated q x q covariance matrices where q = choose(p, 2)}

\item{sigEigs}{List of K matrices containing eigen decomposition matrices for covariance matrices contained in sigmas}

\item{sigMean}{q x q matrix containing element-wise average of sigmas.}

\item{delta}{Threshold for algorithm}

\item{maxIters}{Maximum number of iterations for algorithm}

\item{sig0}{Initial value for sigma parameter}
}
\description{
Variance Components Model
}
\author{
Andrew DiLernia
}
