% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{royVcm_cpp}
\alias{royVcm_cpp}
\title{Variance Components Model using Roy (1989)'s Covariance Estimate}
\usage{
royVcm_cpp(ys, sigmas, sigEigs, delta = 0.001, maxIters = 100L, sig0 = 0.1)
}
\arguments{
\item{ys}{List of K matrices containing observed p-variate time series}

\item{sigmas}{3D array of K estimated q by q covariance matrices where q = choose(p, 2)}

\item{sigEigs}{List of K matrices containing eigen decomposition matrices for covariance matrices contained in sigmas}

\item{delta}{Threshold for algorithm}

\item{maxIters}{Maximum number of iterations for algorithm}

\item{sig0}{Initial value for sigma parameter}
}
\description{
Variance Components Model using Roy (1989)'s Covariance Estimate
}
\author{
Andrew DiLernia
}
