% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{taperCov_cpp}
\alias{taperCov_cpp}
\title{Calculate taper for two vectors}
\usage{
taperCov_cpp(ts1, ts2, banw)
}
\arguments{
\item{ts1}{First time series vector}

\item{ts2}{Second time series vector}

\item{banw}{Bandwidth parameter}
}
\description{
Calculate taper for two vectors
}
\author{
Andrew DiLernia
}
