% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{taperCovSub_cpp}
\alias{taperCovSub_cpp}
\title{Calculate taper for two vectors}
\usage{
taperCovSub_cpp(ts1, ts2, banw, hu2s)
}
\arguments{
\item{ts1}{First time series vector}

\item{ts2}{Second time series vector}

\item{banw}{Bandwidth parameter}

\item{hu2s}{Weights from selected taper function}
}
\description{
Calculate taper for two vectors
}
\author{
Andrew DiLernia
}
