% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{upperTriFill_cpp}
\alias{upperTriFill_cpp}
\title{Calculate upper-triangular matrix}
\usage{
upperTriFill_cpp(n, x)
}
\arguments{
\item{n}{Number of rows / columns of matrix}

\item{x}{Values to fill upper-triangle with}
}
\description{
Calculate upper-triangular matrix
}
\author{
Andrew DiLernia
}
