% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcCovHelper.R
\name{varSim}
\alias{varSim}
\title{Simulate Data from a VAR Model}
\usage{
varSim(nt, coeffMat, covMat)
}
\arguments{
\item{nt}{Number of observations}

\item{coeffMat}{\eqn{p} x \eqn{p} matrix of coefficients}

\item{covMat}{\eqn{p} x \eqn{p} covariance matrix of residuals}
}
\value{
\eqn{nt} x \eqn{p} matrix of the generated \eqn{p}-variate time series
}
\description{
This function simulates data from a mean 0 first-order vector auto-regressive (VAR) model.
}
\examples{
# Generate multivariate time series with 5 variables from a
# first-order VAR model with 50 time points
set.seed(1994)
myTS <- varSim(nt = 50, coeffMat = diag(0.50, 5), covMat = diag(1, 5))

# Plotting generated time series
matplot(myTS, type = 'l', lty = "solid", main = "Generated Multivariate Time Series",
xlab = "Time", ylab = "Value")

}
\author{
Andrew DiLernia
}
