% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdiagArray_cpp}
\alias{bdiagArray_cpp}
\title{Construct Block-Diagonal Matrix}
\usage{
bdiagArray_cpp(array3d)
}
\arguments{
\item{array3d}{3D array of matrices to make into single block-diagonal matrix}
}
\description{
Construct Block-Diagonal Matrix
}
\author{
Andrew DiLernia
}
