% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalFuns.R
\name{multiBlockBoot}
\alias{multiBlockBoot}
\title{Multi-Subject Block-Bootstrap for Partial Correlations of a Multivariate Time Series}
\usage{
multiBlockBoot(mvts, winLength = NULL, nboots = 500)
}
\arguments{
\item{mvts}{List of \eqn{K} number of \eqn{nt} x \eqn{p} matrices of the observed \eqn{p}-variate time series, or \eqn{N} x \eqn{p} x \eqn{K} array.}

\item{winLength}{Specified window length for block-bootstrap (Optional). If not specified, block-length is determined using the method described in Patton, Politis and White (2009).}

\item{nboots}{Number of bootstrap samples.}
}
\value{
A list of length 2 containing:
\enumerate{
\item \eqn{q} x \eqn{q} estimated covariance matrix of the partial correlations where \eqn{q=}choose(\eqn{p}, 2)
\item \eqn{q} x \eqn{2} matrix containing the lower (first column) and upper (second column) bounds of the 95\% bootstrap confidence intervals.
}
}
\description{
This function calculates block-bootstrap covariance matrix estimates for partial correlations of multi-subject time series data.
}
\examples{
# Generate multivariate time series with 5 variables from a
# first-order VAR model with 50 time points
set.seed(1994)
myTS <- varSim(nt = 50, coeffMat = diag(0.50, 5), covMat = diag(1, 5))

# Estimated covariance matrix for partial correlations and 95\% bootstrap intervals
bootRes <- bootVar(myTS)

}
\references{
Politis, D.N. and H. White (2004), Automatic block-length selection for the dependent bootstrap, \emph{Econometric Reviews}, 23(1), 53-70.
}
\author{
Andrew DiLernia
}
