% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{partial_corr_asymptotic_cov_cpp}
\alias{partial_corr_asymptotic_cov_cpp}
\title{Taylor Series Estimate of Covariance Matrix for Partial Correlations}
\usage{
partial_corr_asymptotic_cov_cpp(
  mvts,
  bandwidth,
  q,
  correlation_indices,
  residual_pairs,
  correlation_pairs,
  residuals = NULL,
  diagonal_only = FALSE
)
}
\arguments{
\item{mvts}{\eqn{nt} x \eqn{p} matrix of observed \eqn{p}-variate time series.}

\item{bandwidth}{nonnegative bandwidth parameter.}

\item{q}{number of unique partial correlations equal to choose(\eqn{p}, 2).}

\item{correlation_indices}{matrix of indices for partial correlations equal to unique(royVarhelper(p)[, 1:2]).}

\item{residual_pairs}{matrix of indices for residual pairs equal to royVarhelper(p, errors = TRUE).}

\item{correlation_pairs}{(choose(\eqn{q}, 2) + \eqn{q}) x 4 matrix of indices for partial correlations equal to royVarhelper(p, errors = FALSE).}

\item{diagonal_only}{Logical value, indicating if covariances between correlations are calculated (FALSE) or are returned as 0 (TRUE).}

\item{residuals_matrix}{\eqn{nt} x \eqn{2q} matrix of empirical residuals (optional). If not specified, residuals are obtained using ordinary least squares.}
}
\value{
\eqn{q} x \eqn{q} covariance matrix
}
\description{
This function calculates a second-order Taylor Series estimate of the covariance matrix for partial correlations of a weakly stationary multivariate time series.
}
\author{
Andrew DiLernia
}
