% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{partialCov_cpp}
\alias{partialCov_cpp}
\title{Taylor Series Estimate of Covariance Matrix for Partial Correlations}
\usage{
partialCov_cpp(ts, bw, iMatq, iMate, q)
}
\arguments{
\item{ts}{\eqn{nt} x \eqn{p} matrix of observed p-variate time series.}

\item{bw}{nonnegative bandwidth parameter.}

\item{iMatq}{matrix of indices for partial correlations equal to unique(royVarhelper(p)[, 1:2]).}

\item{iMate}{matrix of indices for partial correlations equal to royVarhelper(p, errors = T).}

\item{\eqn{q}}{number of unique partial correlations equal to choose(\eqn{p}, 2).}
}
\value{
\eqn{q} x \eqn{q} covariance matrix
}
\description{
This function calculates a second-order Taylor Series estimate of the covariance matrix for partial correlations of a weakly stationary multivariate time series.
}
\author{
Andrew DiLernia
}
