% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{royVar_cpp}
\alias{royVar_cpp}
\title{Roy Asymptotic Variance}
\usage{
royVar_cpp(iMat, tsData, q, bw = 10L)
}
\arguments{
\item{iMat}{Matrix of correlation indices}

\item{tsData}{Matrix of observed \eqn{n}-length \eqn{p}-variate time series}

\item{q}{Integer equal to the number of unique variables pairs given by choose(\eqn{p}, 2)}

\item{bw}{Bandwidth parameter}
}
\value{
\eqn{q} x \eqn{q} covariance matrix
}
\description{
This function calculates the asymptotic covariance matrix for correlations of a stationary multivariate time series as derived by Roy (1989).
}
\references{
Roy, R. (1989). Asymptotic covariance structure of serial correlations in multivariate time series. Biometrika, 76(4), 824-827.
}
\author{
Andrew DiLernia
}
